requirements.txt
Django==4.2.6
django-allauth==0.58.0
django-filter==23.3
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Pillow==10.1.0
pydantic==2.4.2
psycopg2-binary==2.9.9
django-ckeditor==6.7.0
redis==5.0.1
celery==5.3.4
# django-crispy-forms==2.1
# crispy-bootstrap4==2023.1
